+++ core/include/sched.h
@@ -203,29 +203,6 @@ extern clist_node_t sched_runqueues[SCHED_PRIO_LEVELS];
  */
 NORETURN void sched_task_exit(void);
 
-#ifdef MODULE_SCHEDSTATISTICS
-/**
- *  Scheduler statistics
- */
-typedef struct {
-    uint32_t laststart;      /**< Time stamp of the last time this thread was
-                                  scheduled to run */
-    unsigned int schedules;  /**< How often the thread was scheduled to run */
-    uint64_t runtime_ticks;  /**< The total runtime of this thread in ticks */
-} schedstat_t;
-
-/**
- *  Thread statistics table
- */
-extern schedstat_t sched_pidlist[KERNEL_PID_LAST + 1];
-
-/**
- *  @brief  Registers the sched statistics callback and sets laststart for
- *          caller thread
- */
-void init_schedstatistics(void);
-#endif /* MODULE_SCHEDSTATISTICS */
-
 #ifdef MODULE_SCHED_CB
 /**
  *  @brief  Register a callback that will be called on every scheduler run
